requirements.txt
numpy==1.18.1
pandas==1.0.1
Pillow==7.2.0
scikit_learn==0.23.1
streamlit==0.62.1
